Added a NewThread library that uses one IO manager per capability.
The NewThread library uses a new manager module to manage only the file descriptors and uses the old manager for managing timeouts. It has only been tested with EPoll; it is likely to need a lot of refinement to make this 1 manager/capability scheme work out well with the other libraries. It changes the semantics of registerFd_: this command now registers the file descriptor to be called back exactly once.