Fix problem with interface between IO managers and RTS.
All signaling from RTS to IO manager is done through the TimerManager. If the timer manager receives a die signal, it shuts down the per-cap managers. Currently, this is done in an indirect way, because the TimerManager doesn't have direct access to the per-cap managers. Instead, it is parameterized with a callback function that it will call when it receives a die message from the RTS. GHC.Event.Thread supplies the callback when it creates the timer manager and the callback it provides performs the shutdown work. To allow the TimerManager to register its signaling file descriptors with the RTS and have the per-cap descriptors not do this, we added a boolean flag to newControl.